Updated the Ink Manager as well as the scene where it was being used.
It should be noted that the Choice UI is a bit weird, as you must select a button using the keyboard and then click again to continue the dialogue, otherwise the dialogue will end prematurely on the Choices.
I don't know why it does that, but I'll figure it out.